Move UserAccessProvider to the Services layer to let it access the old HttpContext api.
