+++ src/Services/Security/UserAccessProvider.cs
@@ -1,0 +1,64 @@
+using SenseNet.ContentRepository.Storage.Security;
+using System.Web;
+using System.Threading;
+
+namespace SenseNet.ContentRepository.Security
+{
+    public class UserAccessProvider : AccessProvider
+    {
+        private IUser CurrentUser
+        {
+            get
+            {
+                if (Thread.CurrentPrincipal.Identity is IUser user)
+                    return user;
+
+                CurrentUser = StartupUser;
+                user = User.Administrator;
+                CurrentUser = user;
+                return user;
+            }
+            set => Thread.CurrentPrincipal = new PortalPrincipal(value);
+        }
+
+        public override IUser GetCurrentUser()
+        {
+            if (HttpContext.Current == null)
+                return CurrentUser;
+
+            IUser currentUser = null;
+            if (HttpContext.Current.User != null)
+                currentUser = HttpContext.Current.User.Identity as IUser;
+
+            if (currentUser != null)
+                return currentUser;
+
+            SetCurrentUser(StartupUser);
+            currentUser = StartupUser;
+
+            return currentUser;
+        }
+
+        protected override void DoSetCurrentUser(IUser user)
+        {
+            if (HttpContext.Current == null)
+                CurrentUser = user;
+            else
+                HttpContext.Current.User = new PortalPrincipal(user);
+        }
+
+        public override bool IsAuthenticated
+        {
+            get
+            {
+                if (HttpContext.Current == null)
+                    return false;
+
+                System.Security.Principal.IPrincipal currentPrincipal =
+                    HttpContext.Current.User;
+
+                return currentPrincipal?.Identity != null && currentPrincipal.Identity.IsAuthenticated;
+            }
+        }
+    }
+}
